Extract shared WindowsGpuStats logic into common base class
Move all GPU metrics business logic (tick aggregation, utilization
calculation, VRAM/shared memory queries, temperature/power/clock/fan
fallback chains) into a common abstract WindowsGpuStats class.

JNA and FFM subclasses implement 18 abstract dispatch methods for
NVML, ADL, PDH (GpuInformation), and LHM WMI calls.

Eliminates ~222 lines of duplicated logic between the two
implementations.

Partial fix for #3274 (Issue 1: JNA/FFM parallel implementations)
